pyproject.toml
[project]
name = "agentic-demo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain",
    "langchain-anthropic",
    "langchain-mcp-adapters",
    "langgraph",
    "fastmcp",
    "httpx",
    "python-dotenv",
]

[dependency-groups]
dev = []
